feat(tiruert): output float instead of decimal numbers

+++ web/tiruert/serializers/teneur.py
@@ -27,7 +27,7 @@ class SimulationInputSerializer(serializers.ModelSerializer):
 
 class SimulationOutputSerializer(serializers.Serializer):
     lot_id = serializers.IntegerField()
-    volume = serializers.DecimalField(max_digits=20, decimal_places=2)
+    volume = serializers.FloatField()
     emission_rate_per_mj = serializers.FloatField()
     fun = serializers.FloatField()
